% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_metrics.R
\name{base_metrics}
\alias{base_metrics}
\title{base_metrics}
\usage{
base_metrics(df, correlation = NULL, pred_name = "pred", obs_name = "obs")
}
\arguments{
\item{df}{Data frame with the predictions and observations as independent columns. The name was chosen to indicate the input is a data frame object}

\item{correlation}{Correlation method to pass cor() function. Accepted values: "pearson", "kendall", "spearman". Default NULL.}

\item{pred_name}{Name of the column containing the predictions. The name was selected to guide the user about what column to select. Default "pred"}

\item{obs_name}{Name of the column containing the observations or prediction targets The name was selected to guide the user about what column to select. Default "obs".}
}
\value{
A dataframe with two columns: first the name of the score calculated, second the score numeric value.
}
\description{
This function calculates three common scores for evaluation of regression models: R-squared, Root Mean Squared Error (RMSE), and a correlation score among "pearson", "kendall", or "spearman". Takes as input a data.frame object with two independent columns for the observed ("obs") and predicted values ("pred").
}
